ref: narrow backend module seams
